add checks to RocksBlockStore

+++ core/server/master/src/main/java/alluxio/master/metastore/rocks/RocksBlockMetaStore.java
@@ -280,6 +280,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public Optional<BlockMeta> getBlock(long id) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     byte[] meta;
     try {
       meta = db().get(mBlockMetaColumn.get(), Longs.toByteArray(id));
@@ -298,6 +301,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public void putBlock(long id, BlockMeta meta) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       byte[] buf = db().get(mBlockMetaColumn.get(), Longs.toByteArray(id));
       // Overwrites the key if it already exists.
@@ -313,6 +319,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public void removeBlock(long id) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     try {
       byte[] buf = db().get(mBlockMetaColumn.get(), Longs.toByteArray(id));
       db().delete(mBlockMetaColumn.get(), mDisableWAL, Longs.toByteArray(id));
@@ -358,6 +367,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public List<BlockLocation> getLocations(long id) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     // References to the RocksObject need to be held explicitly and kept from GC
     // In order to prevent segfaults in the native code execution
     // Ref: https://github.com/facebook/rocksdb/issues/9378
@@ -382,6 +394,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public void addLocation(long id, BlockLocation location) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     byte[] key = RocksUtils.toByteArray(id, location.getWorkerId());
     try {
       db().put(mBlockLocationsColumn.get(), mDisableWAL, key, location.toByteArray());
@@ -392,6 +407,9 @@ public class RocksBlockMetaStore implements BlockMetaStore, RocksCheckpointed {
 
   @Override
   public void removeLocation(long blockId, long workerId) {
+    if (mClosed.get()) {
+      throw new RuntimeException("RocksDB is closed. Master is failing over or shutting down.");
+    }
     byte[] key = RocksUtils.toByteArray(blockId, workerId);
     try {
       db().delete(mBlockLocationsColumn.get(), mDisableWAL, key);
